requirements.txt
# Web framework
streamlit>=1.35.0

# Microsoft Graph / SharePoint
msal>=1.28.0
requests>=2.31.0

# OpenAI
openai>=1.35.0

# LangChain + embeddings + Chroma
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-chroma>=0.1.0
chromadb>=0.5.0
sentence-transformers>=2.2.2

# Document parsing
python-docx>=1.1.0
python-pptx>=0.6.23
PyPDF2>=3.0.1
unstructured>=0.14.5
docx2txt>=0.8
# Optional (advanced PDF parsing, not required by your code):
# pdfminer.six>=20231228

# Utilities
pandas>=2.2.2
numpy>=1.26.4
lxml>=5.2.2
tiktoken>=0.7.0
